requirements_optional_langchain.txt
# To install with constraints
# grep -v '#\|peft' requirements.txt > req_constraints.txt ; pip install -r requirements_optional_langchain.txt -c req_constraints.txt

# optional for chat with PDF
langchain==0.0.171
pypdf==3.8.1
tiktoken==0.3.3
# avoid textract, requires old six
#textract==1.6.5
# choose:
#faiss-cpu
faiss-gpu==1.7.2

# for HF embeddings
sentence_transformers==2.2.2
# for OpenAI embeddings (requires key)
openai==0.27.6

# local vector db
chromadb==0.3.23
# server vector db
#pymilvus==2.2.8

# weak url support, if can't install opencv etc. If comment-in this one, then comment-out unstructured[local-inference]==0.6.6
# unstructured==0.6.6

# strong support for images
# Requires on Ubuntu: sudo apt-get install libmagic-dev poppler-utils tesseract-ocr libreoffice
unstructured[local-inference]==0.6.6
#pdf2image==1.16.3
#pytesseract==0.3.10
pillow

pdfminer.six==20221105
urllib3==1.26.6
requests_file==1.5.1

#pdf2image==1.16.3
#pytesseract==0.3.10
tabulate==0.9.0
# FYI pandoc already part of requirements.txt

jq==1.4.1

# to check licenses
# Run: pip-licenses|grep -v 'BSD\|Apache\|MIT'
pip-licenses==4.3.0
